refactor: 코드 공백 및 테스트코드 정리
