+++ src/main/kotlin/blackjack/domain/RandomCardGenerator.kt
@@ -1,6 +1,6 @@
 package blackjack.domain
 
-class RandomCardGenerator: CardGenerator {
+class RandomCardGenerator : CardGenerator {
     override fun generate(): Card {
         return Card(Rank.random(), Suit.random())
     }
