+++ src/main/kotlin/blackjack/domain/Rank.kt
@@ -7,7 +7,7 @@ enum class Rank(val symbol: String, val score: Int) {
     FOUR("4", 4),
     FIVE("5", 5),
     SIX("6", 6),
-    SECEN("7",7),
+    SECEN("7", 7),
     EIGHT("8", 8),
     NINE("9", 9),
     TEN("10", 10),
