+++ src/test/kotlin/blackjack/fixture/PlayerFixture.kt
@@ -1,8 +1,5 @@
 package blackjack.fixture
 
-import blackjack.domain.Player
-
 object PlayerFixture {
     const val PLAYER_SONG2_NAME = "song2"
-    val PLAYER_SONG2 = Player(PLAYER_SONG2_NAME)
 }
